requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
pymongo==3.12.1
motor==2.5.1
email-validator==1.1.3
starlette==0.16.0
pydantic==1.8.2
python-dotenv==0.19.1
passlib==1.7.4
python-jose==3.3.0
pyJWT==2.3.0








